chore: refactor codebase for artifact hosting
* new Hosting type added to dist-manifest
* dist-manifest merging now generalized
* tasks.rs and lib.rs broken up into more files
* serde annotations cleaned up
