+++ cargo-dist-schema/src/lib.rs
@@ -74,6 +74,10 @@ pub struct DistManifest {
     pub ci: Option<CiInfo>,
     /// Data about dynamic linkage in the built libraries
     pub linkage: Vec<Linkage>,
+    /// Hosting info
+    #[serde(default)]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub hosting: Option<Hosting>,
 }
 
 /// CI backend info
@@ -362,6 +366,7 @@ impl DistManifest {
             publish_prereleases: false,
             ci: None,
             linkage: vec![],
+            hosting: None,
         }
     }
 
@@ -390,6 +395,11 @@ impl DistManifest {
             .artifacts
             .iter()
             .filter_map(|k| Some((&**k, self.artifacts.get(k)?)))
+    }
+
+    /// Get the base URL that artifacts should be downloaded from (append the artifact name to the URL)
+    pub fn artifact_download_url(&self) -> Option<&str> {
+        self.hosting.as_ref()?.live_artifacts_url.as_deref()
     }
 }
 
@@ -420,6 +430,21 @@ pub struct Library {
     /// The package from which a library comes, if relevant
     #[serde(skip_serializing_if = "Option::is_none")]
     pub source: Option<String>,
+}
+
+/// Information about how artifacts are hosted
+#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
+pub struct Hosting {
+    /// A URL you can GET uploaded artifacts from while the release is staged
+    pub staging_artifacts_url: Option<String>,
+    /// A URL you can GET uploaded artifacts from when it's published
+    pub live_artifacts_url: Option<String>,
+    /// A URL you can PUT (upload) artifacts to
+    pub upload_url: Option<String>,
+    /// A URL you can PUT, transitions the release from staged to published
+    pub publish_url: Option<String>,
+    /// A URL you can PUT, announces the release
+    pub announce_url: Option<String>,
 }
 
 /// Helper to read the raw version from serialized json
