+++ cargo-dist/src/lib.rs
@@ -10,41 +10,33 @@
 //! It's currently not terribly well-suited to being used as a pure library because it happily
 //! writes to stderr/stdout whenever it pleases. Suboptimal for a library.
 
-use std::{
-    collections::{BTreeMap, HashMap},
-    env,
-    fs::{self, File},
-    io::{Cursor, Read},
-    process::Command,
-};
-
-use axoasset::{LocalAsset, SourceFile};
+use axoasset::LocalAsset;
 use backend::{
     ci::CiInfo,
-    installer::{self, homebrew::HomebrewInstallerInfo, npm::NpmInstallerInfo, InstallerImpl},
-    templates::{TemplateEntry, TEMPLATE_INSTALLER_NPM},
+    installer::{self, InstallerImpl},
 };
-use camino::{Utf8Path, Utf8PathBuf};
-use cargo_dist_schema::{Asset, AssetKind, DistManifest, ExecutableAsset};
-use comfy_table::{presets::UTF8_FULL, Table};
+use camino::Utf8Path;
+use cargo_build::{build_cargo_target, rustup_toolchain};
+use cargo_dist_schema::DistManifest;
 use config::{
     ArtifactMode, ChecksumStyle, CompressionImpl, Config, DirtyMode, GenerateMode, ZipStyle,
 };
-use goblin::Object;
-use mach_object::{LoadCommand, OFile};
 use semver::Version;
-use serde::{Deserialize, Serialize};
-use tracing::{info, warn};
+use tracing::info;
 
 use errors::*;
 pub use init::{do_init, InitArgs};
-use miette::{miette, Context, IntoDiagnostic};
+use miette::miette;
 pub use tasks::*;
 
+pub mod announce;
 pub mod backend;
+pub mod cargo_build;
 pub mod config;
 pub mod errors;
 mod init;
+pub mod linkage;
+pub mod manifest;
 pub mod tasks;
 #[cfg(test)]
 mod tests;
@@ -53,7 +45,7 @@ mod tests;
 pub fn do_build(cfg: &Config) -> Result<DistManifest> {
     check_integrity(cfg)?;
 
-    let dist = tasks::gather_work(cfg)?;
+    let (dist, mut manifest) = tasks::gather_work(cfg)?;
 
     // FIXME: parallelize this by working this like a dependency graph, so we can start
     // bundling up an executable the moment it's built! Note however that you shouldn't
@@ -75,240 +67,33 @@ pub fn do_build(cfg: &Config) -> Result<DistManifest> {
 
     // Run all the local build steps first
     for step in &dist.local_build_steps {
-        run_build_step(&dist, step, &[])?;
+        run_build_step(&dist, step, &manifest)?;
     }
 
-    // Calculate a temporary build manifest from the output of the local builds
-    // (includes their linkage data)
-    let manifests = vec![build_manifest(cfg, &dist)?];
+    // Compute linkage data now that we're done all builds
+    linkage::add_linkage_to_manifest(cfg, &dist, &mut manifest)?;
 
     // Next the global steps
     for step in &dist.global_build_steps {
-        run_build_step(&dist, step, &manifests)?;
+        run_build_step(&dist, step, &manifest)?;
     }
 
-    Ok(build_manifest(cfg, &dist)?)
+    Ok(manifest)
 }
 
 /// Just generate the manifest produced by `cargo dist build` without building
 pub fn do_manifest(cfg: &Config) -> Result<DistManifest> {
     check_integrity(cfg)?;
-    let dist = gather_work(cfg)?;
-
-    Ok(build_manifest(cfg, &dist)?)
-}
-
-fn build_manifest(cfg: &Config, dist: &DistGraph) -> DistResult<DistManifest> {
-    // Report the releases
-    let mut releases = vec![];
-    let mut all_artifacts = BTreeMap::<String, cargo_dist_schema::Artifact>::new();
-    for release in &dist.releases {
-        // Gather up all the local and global artifacts
-        let mut artifacts = vec![];
-        for &artifact_idx in &release.global_artifacts {
-            let id = &dist.artifact(artifact_idx).id;
-            all_artifacts.insert(id.clone(), manifest_artifact(cfg, dist, artifact_idx));
-            artifacts.push(id.clone());
-        }
-        for &variant_idx in &release.variants {
-            let variant = dist.variant(variant_idx);
-            for &artifact_idx in &variant.local_artifacts {
-                let id = &dist.artifact(artifact_idx).id;
-                all_artifacts.insert(id.clone(), manifest_artifact(cfg, dist, artifact_idx));
-                artifacts.push(id.clone());
-            }
-        }
-
-        // And report the release
-        releases.push(cargo_dist_schema::Release {
-            app_name: release.app_name.clone(),
-            app_version: release.version.to_string(),
-            artifacts,
-        })
-    }
-
-    let linkage = fetch_linkage(
-        cfg.targets.clone(),
-        dist.artifacts.clone(),
-        dist.dist_dir.clone(),
-    )?;
-    let linkage = linkage.iter().map(|l| l.to_schema()).collect();
-
-    let mut manifest = DistManifest::new(releases, all_artifacts);
-
-    // build metadata
-    manifest.dist_version = Some(env!("CARGO_PKG_VERSION").to_owned());
-    manifest.system_info = Some(cargo_dist_schema::SystemInfo {
-        cargo_version_line: dist.tools.cargo.version_line.clone(),
-    });
-
-    // announcement metadata
-    manifest.announcement_tag = dist.announcement_tag.clone();
-    manifest.announcement_is_prerelease = dist.announcement_is_prerelease;
-    manifest.announcement_title = dist.announcement_title.clone();
-    manifest.announcement_changelog = dist.announcement_changelog.clone();
-    manifest.announcement_github_body = dist.announcement_github_body.clone();
-    manifest.system_info = Some(cargo_dist_schema::SystemInfo {
-        cargo_version_line: dist.tools.cargo.version_line.clone(),
-    });
-
-    // ci metadata
-    if !dist.ci_style.is_empty() {
-        let CiInfo { github } = &dist.ci;
-        let github = github.as_ref().map(|info| cargo_dist_schema::GithubCiInfo {
-            artifacts_matrix: Some(info.artifacts_matrix.clone()),
-            pr_run_mode: Some(info.pr_run_mode),
-        });
-
-        manifest.ci = Some(cargo_dist_schema::CiInfo { github });
-    }
-
-    manifest.publish_prereleases = dist.publish_prereleases;
-
-    manifest.linkage = linkage;
+    let (_dist, manifest) = gather_work(cfg)?;
 
     Ok(manifest)
-}
-
-fn manifest_artifact(
-    cfg: &Config,
-    dist: &DistGraph,
-    artifact_idx: ArtifactIdx,
-) -> cargo_dist_schema::Artifact {
-    let artifact = dist.artifact(artifact_idx);
-    let mut assets = vec![];
-
-    let built_assets = artifact
-        .required_binaries
-        .iter()
-        .map(|(&binary_idx, exe_path)| {
-            let binary = &dist.binary(binary_idx);
-            let symbols_artifact = binary.symbols_artifact.map(|a| dist.artifact(a).id.clone());
-            Asset {
-                name: Some(binary.name.clone()),
-                // Always copied to the root... for now
-                path: Some(exe_path.file_name().unwrap().to_owned()),
-                kind: AssetKind::Executable(ExecutableAsset { symbols_artifact }),
-            }
-        });
-
-    let mut static_assets = artifact
-        .archive
-        .as_ref()
-        .map(|archive| {
-            archive
-                .static_assets
-                .iter()
-                .map(|(kind, asset)| {
-                    let kind = match kind {
-                        StaticAssetKind::Changelog => AssetKind::Changelog,
-                        StaticAssetKind::License => AssetKind::License,
-                        StaticAssetKind::Readme => AssetKind::Readme,
-                        StaticAssetKind::Other => AssetKind::Unknown,
-                    };
-                    Asset {
-                        name: Some(asset.file_name().unwrap().to_owned()),
-                        path: Some(asset.file_name().unwrap().to_owned()),
-                        kind,
-                    }
-                })
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-
-    // Record the files that we always add to an npm package
-    //
-    // These can't be pre-included in the normal static assets list above because
-    // they're generated from templates, and not copied from the user's project.
-    if let ArtifactKind::Installer(InstallerImpl::Npm(..)) = &artifact.kind {
-        let root_dir = dist
-            .templates
-            .get_template_dir(TEMPLATE_INSTALLER_NPM)
-            .expect("npm template missing!?");
-        let mut queue = vec![root_dir];
-        while let Some(dir) = queue.pop() {
-            for entry in dir.entries.values() {
-                match entry {
-                    TemplateEntry::Dir(dir) => {
-                        queue.push(dir);
-                    }
-                    TemplateEntry::File(file) => {
-                        static_assets.push(Asset {
-                            name: Some(file.name.clone()),
-                            path: Some(file.path_from_ancestor(root_dir).to_string()),
-                            kind: AssetKind::Unknown,
-                        });
-                    }
-                }
-            }
-        }
-    }
-
-    assets.extend(built_assets);
-    assets.extend(static_assets);
-    // Sort the assets by name to make things extra stable
-    assets.sort_by(|k1, k2| k1.name.cmp(&k2.name));
-
-    let install_hint;
-    let description;
-    let kind;
-
-    match &artifact.kind {
-        ArtifactKind::ExecutableZip(_) => {
-            install_hint = None;
-            description = None;
-            kind = cargo_dist_schema::ArtifactKind::ExecutableZip;
-        }
-        ArtifactKind::Symbols(_) => {
-            install_hint = None;
-            description = None;
-            kind = cargo_dist_schema::ArtifactKind::Symbols;
-        }
-        ArtifactKind::Installer(
-            InstallerImpl::Powershell(info)
-            | InstallerImpl::Shell(info)
-            | InstallerImpl::Homebrew(HomebrewInstallerInfo { inner: info, .. })
-            | InstallerImpl::Npm(NpmInstallerInfo { inner: info, .. }),
-        ) => {
-            install_hint = Some(info.hint.clone());
-            description = Some(info.desc.clone());
-            kind = cargo_dist_schema::ArtifactKind::Installer;
-        }
-        ArtifactKind::Installer(InstallerImpl::Msi(..)) => {
-            install_hint = None;
-            description = Some("install via msi".to_owned());
-            kind = cargo_dist_schema::ArtifactKind::Installer;
-        }
-        ArtifactKind::Checksum(_) => {
-            install_hint = None;
-            description = None;
-            kind = cargo_dist_schema::ArtifactKind::Checksum;
-        }
-    };
-
-    let checksum = artifact.checksum.map(|idx| dist.artifact(idx).id.clone());
-
-    cargo_dist_schema::Artifact {
-        name: Some(artifact.id.clone()),
-        path: if cfg.no_local_paths {
-            None
-        } else {
-            Some(artifact.file_path.to_string())
-        },
-        target_triples: artifact.target_triples.clone(),
-        install_hint,
-        description,
-        assets,
-        kind,
-        checksum,
-    }
 }
 
 /// Run some build step
 fn run_build_step(
     dist_graph: &DistGraph,
     target: &BuildStep,
-    manifests: &[DistManifest],
+    manifest: &DistManifest,
 ) -> Result<()> {
     match target {
         BuildStep::Cargo(target) => build_cargo_target(dist_graph, target),
@@ -332,7 +117,7 @@ fn run_build_step(
             with_root,
         }) => zip_dir(src_path, dest_path, zip_style, with_root.as_deref()),
         BuildStep::GenerateInstaller(installer) => {
-            generate_installer(dist_graph, installer, manifests)
+            generate_installer(dist_graph, installer, manifest)
         }
         BuildStep::Checksum(ChecksumImpl {
             checksum,
@@ -397,333 +182,6 @@ fn write_checksum(checksum: &str, src_path: &Utf8Path, dest_path: &Utf8Path) -> 
     let file_path = src_path.file_name().expect("hashing file with no name!?");
     let line = format!("{checksum} *{file_path}\n");
     axoasset::LocalAsset::write_new(&line, dest_path)?;
-    Ok(())
-}
-
-// Takes a string in KEY=value environment variable format and
-// parses it into a BTreeMap. The string syntax is sh-compatible, and also the
-// format returned by `env`.
-// Note that we trust the parsed string to contain a given key only once;
-// if specified more than once, only the final occurrence will be included.
-fn parse_env(env_string: &str) -> DistResult<BTreeMap<&str, &str>> {
-    let mut parsed = BTreeMap::new();
-    for line in env_string.trim_end().split('\n') {
-        let Some((key, value)) = line.split_once('=') else {
-            return Err(DistError::EnvParseError {
-                line: line.to_owned(),
-            });
-        };
-        parsed.insert(key, value);
-    }
-
-    Ok(parsed)
-}
-
-/// Given the environment captured from `brew bundle exec -- env`, returns
-/// a list of all dependencies from that environment and the opt prefixes
-/// to those packages.
-fn formulas_from_env(environment: &BTreeMap<&str, &str>) -> Vec<(String, String)> {
-    let mut packages = vec![];
-
-    // Set by Homebrew/brew bundle - a comma-separated list of all
-    // dependencies in the recursive tree calculated from the dependencies
-    // in the Brewfile.
-    if let Some(formulastring) = environment.get("HOMEBREW_DEPENDENCIES") {
-        // Set by Homebrew/brew bundle - the path to Homebrew's "opt"
-        // directory, which is where links to the private cellar of every
-        // installed package lives.
-        // Usually /opt/homebrew/opt or /usr/local/opt.
-        if let Some(opt_prefix) = environment.get("HOMEBREW_OPT") {
-            for dep in formulastring.split(',') {
-                // Unwrap here is safe because `split` will always return
-                // a collection of at least one item.
-                let short_name = dep.split('/').last().unwrap();
-                let pkg_opt = format!("{opt_prefix}/{short_name}");
-                packages.push((dep.to_owned(), pkg_opt));
-            }
-        }
-    }
-
-    packages
-}
-
-/// Takes a BTreeMap of key/value environment variables produced by
-/// `brew bundle exec` and decides which ones we want to keep for our own builds.
-/// Returns a Vec containing (KEY, value) tuples.
-fn select_brew_env(environment: &BTreeMap<&str, &str>) -> Vec<(String, String)> {
-    let mut desired_env = vec![];
-
-    // Several of Homebrew's environment variables are safe for us to use
-    // unconditionally, so pick those in their entirety.
-    if let Some(value) = environment.get("PKG_CONFIG_PATH") {
-        desired_env.push(("PKG_CONFIG_PATH".to_owned(), value.to_string()))
-    }
-    if let Some(value) = environment.get("PKG_CONFIG_LIBDIR") {
-        desired_env.push(("PKG_CONFIG_LIBDIR".to_owned(), value.to_string()))
-    }
-    if let Some(value) = environment.get("CMAKE_INCLUDE_PATH") {
-        desired_env.push(("CMAKE_INCLUDE_PATH".to_owned(), value.to_string()))
-    }
-    if let Some(value) = environment.get("CMAKE_LIBRARY_PATH") {
-        desired_env.push(("CMAKE_LIBRARY_PATH".to_owned(), value.to_string()))
-    }
-    let mut paths = vec![];
-
-    // For each listed dependency, add it to the PATH
-    for (_, pkg_opt) in formulas_from_env(environment) {
-        // Not every package will have a /bin or /sbin directory,
-        // but it's safe to add both to the PATH just in case.
-        paths.push(format!("{pkg_opt}/bin"));
-        paths.push(format!("{pkg_opt}/sbin"));
-    }
-
-    if !paths.is_empty() {
-        let our_path = env!("PATH");
-        let desired_path = format!("{our_path}:{}", paths.join(":"));
-
-        desired_env.insert(0, ("PATH".to_owned(), desired_path));
-    }
-
-    desired_env
-}
-
-/// Similar to the above, we read Homebrew's recursive dependency tree and
-/// then append link flags to cargo-dist's rustflags.
-/// These ensure that Rust can find C libraries that may exist within
-/// each package's prefix.
-fn determine_brew_rustflags(base_rustflags: &str, environment: &BTreeMap<&str, &str>) -> String {
-    let mut rustflags = base_rustflags.to_owned();
-    // For each listed dependency, add it to CFLAGS/LDFLAGS
-    for (_, pkg_opt) in formulas_from_env(environment) {
-        // Note that this path might not actually exist; not every
-        // package contains libraries. However, it's safe to
-        // append this flag anyway; Rust passes it on to the
-        // compiler/linker, which tolerate missing directories
-        // just fine.
-        rustflags = format!("{rustflags} -L{pkg_opt}/lib");
-    }
-
-    rustflags
-}
-
-/// Build a cargo target
-fn build_cargo_target(dist_graph: &DistGraph, target: &CargoBuildStep) -> Result<()> {
-    eprint!(
-        "building cargo target ({}/{}",
-        target.target_triple, target.profile
-    );
-
-    let mut rustflags = target.rustflags.clone();
-    let mut desired_extra_env = vec![];
-    let skip_brewfile = env::var("DO_NOT_USE_BREWFILE").is_ok();
-    if let Some(brew) = &dist_graph.tools.brew {
-        if Utf8Path::new("Brewfile").exists() && !skip_brewfile {
-            // Uses `brew bundle exec` to just print its own environment,
-            // allowing us to capture what it generated and decide what
-            // to do with it.
-            let result = Command::new(&brew.cmd)
-                .arg("bundle")
-                .arg("exec")
-                .arg("--")
-                .arg("/usr/bin/env")
-                .output()
-                .into_diagnostic()
-                .wrap_err_with(|| "failed to exec brew bundle exec".to_string())?;
-
-            let env_output = String::from_utf8_lossy(&result.stdout).to_string();
-
-            let brew_env = parse_env(&env_output)?;
-            desired_extra_env = select_brew_env(&brew_env);
-            rustflags = determine_brew_rustflags(&rustflags, &brew_env);
-        }
-    }
-
-    let mut command = Command::new(&dist_graph.tools.cargo.cmd);
-    command
-        .arg("build")
-        .arg("--profile")
-        .arg(&target.profile)
-        .arg("--message-format=json-render-diagnostics")
-        .arg("--target")
-        .arg(&target.target_triple)
-        .env("RUSTFLAGS", &rustflags)
-        .stdout(std::process::Stdio::piped());
-    if !target.features.default_features {
-        command.arg("--no-default-features");
-    }
-    match &target.features.features {
-        CargoTargetFeatureList::All => {
-            command.arg("--all-features");
-        }
-        CargoTargetFeatureList::List(features) => {
-            if !features.is_empty() {
-                // The way we pass these, Cargo wants us to use --features
-                // once for each arg, idk why exactly (might be a windows quirk).
-                for feature in features {
-                    command.arg("--features");
-                    command.arg(feature);
-                }
-            }
-        }
-    }
-    match &target.package {
-        CargoTargetPackages::Workspace => {
-            command.arg("--workspace");
-            eprintln!(" --workspace)");
-        }
-        CargoTargetPackages::Package(package) => {
-            command.arg("--package").arg(package);
-            eprintln!(" --package={})", package);
-        }
-    }
-    // If we generated any extra environment variables to
-    // inject into the environment, apply them now.
-    command.envs(desired_extra_env);
-    info!("exec: {:?}", command);
-    let mut task = command
-        .spawn()
-        .into_diagnostic()
-        .wrap_err_with(|| format!("failed to exec cargo build: {command:?}"))?;
-
-    // Create entries for all the binaries we expect to find, and the paths they should
-    // be copied to (according to the copy_exe_to subscribers list).
-    //
-    // Structure is:
-    //
-    // package-id (key)
-    //    binary-name (key)
-    //       subscribers (list)
-    //          src-path (initially blank, must be filled in by rustc)
-    //          dest-path (where to copy the file to)
-    let mut expected_exes =
-        HashMap::<String, HashMap<String, Vec<(Utf8PathBuf, Utf8PathBuf)>>>::new();
-    let mut expected_symbols =
-        HashMap::<String, HashMap<String, Vec<(Utf8PathBuf, Utf8PathBuf)>>>::new();
-    for &binary_idx in &target.expected_binaries {
-        let binary = &dist_graph.binary(binary_idx);
-        let package_id = binary.pkg_id.to_string();
-        let exe_name = binary.name.clone();
-        for exe_dest in &binary.copy_exe_to {
-            expected_exes
-                .entry(package_id.clone())
-                .or_default()
-                .entry(exe_name.clone())
-                .or_default()
-                .push((Utf8PathBuf::new(), exe_dest.clone()));
-        }
-        for sym_dest in &binary.copy_symbols_to {
-            expected_symbols
-                .entry(package_id.clone())
-                .or_default()
-                .entry(exe_name.clone())
-                .or_default()
-                .push((Utf8PathBuf::new(), sym_dest.clone()));
-        }
-    }
-
-    // Collect up the compiler messages to find out where binaries ended up
-    let reader = std::io::BufReader::new(task.stdout.take().unwrap());
-    for message in cargo_metadata::Message::parse_stream(reader) {
-        let Ok(message) = message
-            .into_diagnostic()
-            .wrap_err("failed to parse cargo json message")
-            .map_err(|e| warn!("{:?}", e))
-        else {
-            // It's ok for there to be messages we don't understand if we don't care about them.
-            // At the end we'll check if we got the messages we *do* need.
-            continue;
-        };
-        match message {
-            cargo_metadata::Message::CompilerArtifact(artifact) => {
-                // Hey we got an executable, is it one we wanted?
-                if let Some(new_exe) = artifact.executable {
-                    info!("got a new exe: {}", new_exe);
-                    let package_id = artifact.package_id.to_string();
-                    let exe_name = new_exe.file_stem().unwrap();
-
-                    // If we expected some symbols, pull them out of the paths of this executable
-                    let expected_sym = expected_symbols
-                        .get_mut(&package_id)
-                        .and_then(|m| m.get_mut(exe_name));
-                    if let Some(expected) = expected_sym {
-                        for (src_sym_path, _) in expected {
-                            for path in &artifact.filenames {
-                                // FIXME: unhardcode this when we add support for other symbol kinds!
-                                let is_symbols =
-                                    path.extension().map(|e| e == "pdb").unwrap_or(false);
-                                if is_symbols {
-                                    // These are symbols we expected! Save the path.
-                                    *src_sym_path = path.to_owned();
-                                }
-                            }
-                        }
-                    }
-
-                    // Get the exe path
-                    let expected_exe = expected_exes
-                        .get_mut(&package_id)
-                        .and_then(|m| m.get_mut(exe_name));
-                    if let Some(expected) = expected_exe {
-                        for (src_bin_path, _) in expected {
-                            // This is an exe we expected! Save the path.
-                            *src_bin_path = new_exe.clone();
-                        }
-                    }
-                }
-            }
-            _ => {
-                // Nothing else interesting?
-            }
-        }
-    }
-
-    // Check that we got everything we expected, and normalize to ArtifactIdx => Artifact Path
-    for (package_id, exes) in expected_exes {
-        for (exe_name, to_copy) in &exes {
-            for (src_path, dest_path) in to_copy {
-                if src_path.as_str().is_empty() {
-                    return Err(miette!(
-                        "failed to find bin {} ({}) -- did the cargo build above have errors?",
-                        exe_name,
-                        package_id
-                    ));
-                }
-                copy_file(src_path, dest_path)?;
-            }
-        }
-    }
-    for (package_id, symbols) in expected_symbols {
-        for (exe, to_copy) in &symbols {
-            for (src_path, dest_path) in to_copy {
-                if src_path.as_str().is_empty() {
-                    return Err(miette!(
-                        "failed to find symbols for bin {} ({}) -- did the cargo build above have errors?",
-                        exe,
-                        package_id
-                    ));
-                }
-                copy_file(src_path, dest_path)?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// Build a cargo target
-fn rustup_toolchain(_dist_graph: &DistGraph, cmd: &RustupStep) -> Result<()> {
-    eprintln!("running rustup to ensure you have {} installed", cmd.target);
-    let status = Command::new(&cmd.rustup.cmd)
-        .arg("target")
-        .arg("add")
-        .arg(&cmd.target)
-        .status()
-        .into_diagnostic()
-        .wrap_err("Failed to install rustup toolchain")?;
-
-    if !status.success() {
-        return Err(miette!("Failed to install rustup toolchain"));
-    }
     Ok(())
 }
 
@@ -800,17 +258,6 @@ pub struct GenerateArgs {
     pub modes: Vec<GenerateMode>,
 }
 
-/// Arguments for `cargo dist linkage` ([`do_linkage][])
-#[derive(Debug)]
-pub struct LinkageArgs {
-    /// Print human-readable output
-    pub print_output: bool,
-    /// Print output as JSON
-    pub print_json: bool,
-    /// Read linkage data from JSON rather than performing a live check
-    pub from_json: Option<String>,
-}
-
 fn do_generate_preflight_checks(dist: &DistGraph) -> Result<()> {
     // Enforce cargo-dist-version, unless...
     //
@@ -839,7 +286,7 @@ fn do_generate_preflight_checks(dist: &DistGraph) -> Result<()> {
 
 /// Generate any scripts which are relevant (impl of `cargo dist generate`)
 pub fn do_generate(cfg: &Config, args: &GenerateArgs) -> Result<()> {
-    let dist = gather_work(cfg)?;
+    let (dist, _manifest) = gather_work(cfg)?;
 
     run_generate(&dist, args)?;
 
@@ -903,463 +350,6 @@ pub fn run_generate(dist: &DistGraph, args: &GenerateArgs) -> Result<()> {
     Ok(())
 }
 
-fn fetch_linkage(
-    targets: Vec<String>,
-    artifacts: Vec<Artifact>,
-    dist_dir: Utf8PathBuf,
-) -> DistResult<Vec<Linkage>> {
-    let mut reports = vec![];
-
-    for target in targets {
-        let artifacts: Vec<Artifact> = artifacts
-            .clone()
-            .into_iter()
-            .filter(|r| r.target_triples.contains(&target))
-            .collect();
-
-        if artifacts.is_empty() {
-            eprintln!("No matching artifact for target {target}");
-            continue;
-        }
-
-        for artifact in artifacts {
-            let path = Utf8PathBuf::from(&dist_dir).join(format!("{}-{target}", artifact.id));
-
-            for (_, binary) in artifact.required_binaries {
-                let bin_path = path.join(binary);
-                if !bin_path.exists() {
-                    eprintln!("Binary {bin_path} missing; skipping check");
-                } else {
-                    reports.push(determine_linkage(&bin_path, &target)?);
-                }
-            }
-        }
-    }
-
-    Ok(reports)
-}
-
-/// Determinage dynamic linkage of built artifacts (impl of `cargo dist linkage`)
-pub fn do_linkage(cfg: &Config, args: &LinkageArgs) -> Result<()> {
-    let dist = gather_work(cfg)?;
-
-    let reports: Vec<Linkage> = if let Some(target) = args.from_json.clone() {
-        let file = SourceFile::load_local(&target)?;
-        file.deserialize_json()?
-    } else {
-        fetch_linkage(cfg.targets.clone(), dist.artifacts, dist.dist_dir)?
-    };
-
-    if args.print_output {
-        for report in &reports {
-            eprintln!("{}", report.report());
-        }
-    }
-    if args.print_json {
-        let j = serde_json::to_string(&reports).unwrap();
-        println!("{}", j);
-    }
-
-    Ok(())
-}
-
-/// Information about dynamic libraries used by a binary
-#[derive(Debug, Deserialize, Serialize)]
-pub struct Linkage {
-    /// The filename of the binary
-    pub binary: String,
-    /// The target triple for which the binary was built
-    pub target: String,
-    /// Libraries included with the operating system
-    pub system: Vec<Library>,
-    /// Libraries provided by the Homebrew package manager
-    pub homebrew: Vec<Library>,
-    /// Public libraries not provided by the system and not managed by any package manager
-    pub public_unmanaged: Vec<Library>,
-    /// Libraries which don't fall into any other categories
-    pub other: Vec<Library>,
-    /// Frameworks, only used on macOS
-    pub frameworks: Vec<Library>,
-}
-
-impl Linkage {
-    /// Formatted human-readable output
-    pub fn report(&self) -> String {
-        let mut table = Table::new();
-        table
-            .load_preset(UTF8_FULL)
-            .set_header(vec!["Category", "Libraries"])
-            .add_row(vec![
-                "System",
-                self.system
-                    .clone()
-                    .into_iter()
-                    .map(|l| l.to_string_pretty())
-                    .collect::<Vec<String>>()
-                    .join("\n")
-                    .as_str(),
-            ])
-            .add_row(vec![
-                "Homebrew",
-                self.homebrew
-                    .clone()
-                    .into_iter()
-                    .map(|l| l.to_string_pretty())
-                    .collect::<Vec<String>>()
-                    .join("\n")
-                    .as_str(),
-            ])
-            .add_row(vec![
-                "Public (unmanaged)",
-                self.public_unmanaged
-                    .clone()
-                    .into_iter()
-                    .map(|l| l.path)
-                    .collect::<Vec<String>>()
-                    .join("\n")
-                    .as_str(),
-            ])
-            .add_row(vec![
-                "Frameworks",
-                self.frameworks
-                    .clone()
-                    .into_iter()
-                    .map(|l| l.path)
-                    .collect::<Vec<String>>()
-                    .join("\n")
-                    .as_str(),
-            ])
-            .add_row(vec![
-                "Other",
-                self.other
-                    .clone()
-                    .into_iter()
-                    .map(|l| l.to_string_pretty())
-                    .collect::<Vec<String>>()
-                    .join("\n")
-                    .as_str(),
-            ]);
-
-        let s = format!(
-            r#"{} ({}):
-
-{table}"#,
-            self.binary, self.target,
-        );
-
-        s.to_owned()
-    }
-
-    fn to_schema(&self) -> cargo_dist_schema::Linkage {
-        cargo_dist_schema::Linkage {
-            binary: self.binary.clone(),
-            target: self.target.clone(),
-            system: self.system.iter().map(|s| s.to_schema()).collect(),
-            homebrew: self.homebrew.iter().map(|s| s.to_schema()).collect(),
-            public_unmanaged: self
-                .public_unmanaged
-                .iter()
-                .map(|s| s.to_schema())
-                .collect(),
-            other: self.other.iter().map(|s| s.to_schema()).collect(),
-            frameworks: self.frameworks.iter().map(|s| s.to_schema()).collect(),
-        }
-    }
-
-    /// Constructs a Linkage from a cargo_dist_schema::Linkage
-    pub fn from_schema(other: &cargo_dist_schema::Linkage) -> Self {
-        Self {
-            binary: other.binary.clone(),
-            target: other.target.clone(),
-            system: other.system.iter().map(Library::from_schema).collect(),
-            homebrew: other.homebrew.iter().map(Library::from_schema).collect(),
-            public_unmanaged: other
-                .public_unmanaged
-                .iter()
-                .map(Library::from_schema)
-                .collect(),
-            other: other.other.iter().map(Library::from_schema).collect(),
-            frameworks: other.frameworks.iter().map(Library::from_schema).collect(),
-        }
-    }
-}
-
-/// Represents a dynamic library located somewhere on the system
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct Library {
-    /// The path to the library; on platforms without that information, it will be a basename instead
-    pub path: String,
-    /// The package from which a library comes, if relevant
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-}
-
-impl Library {
-    fn new(library: String) -> Self {
-        Self {
-            path: library,
-            source: None,
-        }
-    }
-
-    fn to_schema(&self) -> cargo_dist_schema::Library {
-        cargo_dist_schema::Library {
-            path: self.path.clone(),
-            source: self.source.clone(),
-        }
-    }
-
-    fn from_schema(other: &cargo_dist_schema::Library) -> Self {
-        Self {
-            path: other.path.clone(),
-            source: other.source.clone(),
-        }
-    }
-
-    fn from_homebrew(library: String) -> Self {
-        // Doesn't currently support Homebrew installations in
-        // non-default locations
-        let brew_prefix = if library.starts_with("/opt/homebrew/opt/") {
-            Some("/opt/homebrew/opt/")
-        } else if library.starts_with("/usr/local/opt/") {
-            Some("/usr/local/opt/")
-        } else {
-            None
-        };
-
-        if let Some(prefix) = brew_prefix {
-            let cloned = library.clone();
-            let stripped = cloned.strip_prefix(prefix).unwrap();
-            let mut package = stripped.split('/').nth(0).unwrap().to_owned();
-
-            // The path alone isn't enough to determine the tap the formula
-            // came from. If the install receipt exists, we can use it to
-            // get the name of the source tap.
-            let receipt = Utf8PathBuf::from(&prefix)
-                .join(&package)
-                .join("INSTALL_RECEIPT.json");
-
-            // If the receipt doesn't exist or can't be loaded, that's not an
-            // error; we can fall back to the package basename we parsed out
-            // of the path.
-            if receipt.exists() {
-                let _ = SourceFile::load_local(&receipt)
-                    .and_then(|file| file.deserialize_json())
-                    .map(|parsed: serde_json::Value| {
-                        if let Some(tap) = parsed["source"]["tap"].as_str() {
-                            if tap != "homebrew/core" {
-                                package = format!("{tap}/{package}");
-                            }
-                        }
-                    });
-            }
-
-            Self {
-                path: library,
-                source: Some(package.to_owned()),
-            }
-        } else {
-            Self {
-                path: library,
-                source: None,
-            }
-        }
-    }
-
-    fn maybe_apt(library: String) -> DistResult<Self> {
-        // We can't get this information on other OSs
-        if std::env::consts::OS != "linux" {
-            return Ok(Self {
-                path: library,
-                source: None,
-            });
-        }
-
-        let process = Command::new("dpkg")
-            .arg("--search")
-            .arg(&library)
-            .output()
-            .into_diagnostic();
-        match process {
-            Ok(output) => {
-                let output = String::from_utf8(output.stdout)?;
-
-                let package = output.split(':').nth(0).unwrap();
-                let source = if package.is_empty() {
-                    None
-                } else {
-                    Some(package.to_owned())
-                };
-
-                Ok(Self {
-                    path: library,
-                    source,
-                })
-            }
-            // Couldn't find a package for this file
-            Err(_) => Ok(Self {
-                path: library,
-                source: None,
-            }),
-        }
-    }
-
-    fn to_string_pretty(&self) -> String {
-        if let Some(package) = &self.source {
-            format!("{} ({package})", self.path).to_owned()
-        } else {
-            self.path.clone()
-        }
-    }
-}
-
-fn do_otool(path: &Utf8PathBuf) -> DistResult<Vec<String>> {
-    let mut libraries = vec![];
-
-    let mut f = File::open(path)?;
-    let mut buf = vec![];
-    let size = f.read_to_end(&mut buf).unwrap();
-    let mut cur = Cursor::new(&buf[..size]);
-    if let OFile::MachFile {
-        header: _,
-        commands,
-    } = OFile::parse(&mut cur).unwrap()
-    {
-        let commands = commands
-            .iter()
-            .map(|load| load.command())
-            .cloned()
-            .collect::<Vec<LoadCommand>>();
-
-        for command in commands {
-            match command {
-                LoadCommand::IdDyLib(ref dylib)
-                | LoadCommand::LoadDyLib(ref dylib)
-                | LoadCommand::LoadWeakDyLib(ref dylib)
-                | LoadCommand::ReexportDyLib(ref dylib)
-                | LoadCommand::LoadUpwardDylib(ref dylib)
-                | LoadCommand::LazyLoadDylib(ref dylib) => {
-                    libraries.push(dylib.name.to_string());
-                }
-                _ => {}
-            }
-        }
-    }
-
-    Ok(libraries)
-}
-
-fn do_ldd(path: &Utf8PathBuf) -> DistResult<Vec<String>> {
-    let mut libraries = vec![];
-
-    let output = Command::new("ldd")
-        .arg(path)
-        .output()
-        .expect("Unable to run ldd");
-
-    let result = String::from_utf8_lossy(&output.stdout).to_string();
-    let lines = result.trim_end().split('\n');
-
-    for line in lines {
-        let line = line.trim();
-
-        // There's no dynamic linkage at all; we can safely break,
-        // there will be nothing useful to us here.
-        if line.starts_with("not a dynamic executable") {
-            break;
-        }
-
-        // Not a library that actually concerns us
-        if line.starts_with("linux-vdso") {
-            continue;
-        }
-
-        // Format: libname.so.1 => /path/to/libname.so.1 (address)
-        if let Some(path) = line.split(" => ").nth(1) {
-            // This may be a symlink rather than the actual underlying library;
-            // we resolve the symlink here so that we return the real paths,
-            // making it easier to map them to their packages later.
-            let lib = (path.split(' ').next().unwrap()).to_owned();
-            let realpath = fs::canonicalize(&lib)?;
-            libraries.push(realpath.to_string_lossy().to_string());
-        } else {
-            continue;
-        }
-    }
-
-    Ok(libraries)
-}
-
-fn do_pe(path: &Utf8PathBuf) -> DistResult<Vec<String>> {
-    let buf = std::fs::read(path)?;
-    match Object::parse(&buf)? {
-        Object::PE(pe) => Ok(pe.libraries.into_iter().map(|s| s.to_owned()).collect()),
-        _ => Err(DistError::LinkageCheckUnsupportedBinary {}),
-    }
-}
-
-fn determine_linkage(path: &Utf8PathBuf, target: &str) -> DistResult<Linkage> {
-    let libraries = match target {
-        // Can be run on any OS
-        "i686-apple-darwin" | "x86_64-apple-darwin" | "aarch64-apple-darwin" => do_otool(path)?,
-        "i686-unknown-linux-gnu"
-        | "x86_64-unknown-linux-gnu"
-        | "aarch64-unknown-linux-gnu"
-        | "i686-unknown-linux-musl"
-        | "x86_64-unknown-linux-musl"
-        | "aarch64-unknown-linux-musl" => {
-            // Currently can only be run on Linux
-            if std::env::consts::OS != "linux" {
-                return Err(DistError::LinkageCheckInvalidOS {
-                    host: std::env::consts::OS.to_owned(),
-                    target: target.to_owned(),
-                });
-            }
-            do_ldd(path)?
-        }
-        // Can be run on any OS
-        "i686-pc-windows-msvc" | "x86_64-pc-windows-msvc" | "aarch64-pc-windows-msvc" => {
-            do_pe(path)?
-        }
-        _ => return Err(DistError::LinkageCheckUnsupportedBinary {}),
-    };
-
-    let mut linkage = Linkage {
-        binary: path.file_name().unwrap().to_owned(),
-        target: target.to_owned(),
-        system: vec![],
-        homebrew: vec![],
-        public_unmanaged: vec![],
-        frameworks: vec![],
-        other: vec![],
-    };
-    for library in libraries {
-        if library.starts_with("/opt/homebrew") {
-            linkage
-                .homebrew
-                .push(Library::from_homebrew(library.clone()));
-        } else if library.starts_with("/usr/lib") || library.starts_with("/lib") {
-            linkage.system.push(Library::maybe_apt(library.clone())?);
-        } else if library.starts_with("/System/Library/Frameworks")
-            || library.starts_with("/Library/Frameworks")
-        {
-            linkage.frameworks.push(Library::new(library.clone()));
-        } else if library.starts_with("/usr/local") {
-            if std::fs::canonicalize(&library)?.starts_with("/usr/local/Cellar") {
-                linkage
-                    .homebrew
-                    .push(Library::from_homebrew(library.clone()));
-            } else {
-                linkage.public_unmanaged.push(Library::new(library.clone()));
-            }
-        } else {
-            linkage.other.push(Library::maybe_apt(library.clone())?);
-        }
-    }
-
-    Ok(linkage)
-}
-
 /// Run any necessary integrity checks for "primary" commands like build/plan
 ///
 /// (This is currently equivalent to `cargo dist generate --check`)
@@ -1376,7 +366,7 @@ pub fn check_integrity(cfg: &Config) -> Result<()> {
         installers: vec![],
         announcement_tag: None,
     };
-    let dist = tasks::gather_work(&check_config)?;
+    let (dist, _manifest) = tasks::gather_work(&check_config)?;
 
     run_generate(
         &dist,
@@ -1391,7 +381,7 @@ pub fn check_integrity(cfg: &Config) -> Result<()> {
 fn generate_installer(
     dist: &DistGraph,
     style: &InstallerImpl,
-    manifests: &[DistManifest],
+    manifest: &DistManifest,
 ) -> Result<()> {
     match style {
         InstallerImpl::Shell(info) => {
@@ -1402,7 +392,7 @@ fn generate_installer(
         }
         InstallerImpl::Npm(info) => installer::npm::write_npm_project(&dist.templates, info)?,
         InstallerImpl::Homebrew(info) => {
-            installer::homebrew::write_homebrew_formula(&dist.templates, dist, info, manifests)?
+            installer::homebrew::write_homebrew_formula(&dist.templates, dist, info, manifest)?
         }
         InstallerImpl::Msi(info) => info.build()?,
     }
